移除多余的逻辑删除抽象
