+++ mybatis-plus-generate/src/main/java/com/baomidou/mybatisplus/generator/config/StrategyConfig.java
@@ -432,8 +432,9 @@ public class StrategyConfig {
         return fieldPrefix;
     }
 
-    public void setFieldPrefix(String[] fieldPrefix) {
+    public StrategyConfig setFieldPrefix(String[] fieldPrefix) {
         this.fieldPrefix = fieldPrefix;
+        return this;
     }
 
     public StrategyConfig fieldPrefix(String... fieldPrefixs) {
